--- a/src/main/java/com/iota/iri/crypto/Curl.java
+++ b/src/main/java/com/iota/iri/crypto/Curl.java
@@ -136,10 +136,9 @@ public class Curl implements Sponge {
             for (int curlStateIndex = 0; curlStateIndex < STATE_LENGTH; curlStateIndex++) {
                 final long alpha = curlScratchpadLow[curlScratchpadIndex];
                 final long beta = curlScratchpadHigh[curlScratchpadIndex];
-                final long gamma = curlScratchpadHigh[curlScratchpadIndex += (curlScratchpadIndex < 365 ? 364 : -365)];
-                final long delta = (alpha | (~gamma)) & (curlScratchpadLow[curlScratchpadIndex] ^ beta);
-                stateLow[curlStateIndex] = ~delta;
-                stateHigh[curlStateIndex] = (alpha ^ gamma) | delta;
+                final long delta = beta ^ curlScratchpadLow[curlScratchpadIndex += (curlScratchpadIndex < 365 ? 364 : -365)];
+                stateLow[curlStateIndex] = ~(delta & alpha);
+                stateHigh[curlStateIndex] = delta | (alpha ^ curlScratchpadHigh[curlScratchpadIndex]);
             }
         }
     }