# Description

powsrv.io Team optimized the curl transform formulas.
On my machine it's around 7% faster than before in golang and in C.
Didn't benchmark the improvement in Java :)

## Type of change

- Enhancement (a non-breaking change which adds functionality)


# How Has This Been Tested?

- PearlDiverTest.java
- CurlTest.java


# Checklist:

- [X] My code follows the style guidelines for this project
- [X] I have performed a self-review of my own code
- [X] New and existing unit tests pass locally with my changes