@GalRogozinski 2019-02-10
@nmeylan 
Did you run the test in a clean slate where nothing runs in the background?
The results are not very encouraging :-(

By the way the benchmark code itself lgtm :-)
If you want you can PR it. We have a benchmarks folder in IRI.
Just please add a new line above `@Benchmark` and make `pairHashWorks` private (not so important it just hurts my eyes)