@nmeylan 2019-02-10
@GalRogozinski 
Unfortunately, I didn't run this test in an isolated environment, but both executions ran in same condition.

I've also edited my comment to use **Throughput** instead of **Average Time** which gives clearer results.