@GalRogozinski 2019-02-10
Hmm I see now the PR has higher throughput which is good.
But before your edit it had a higher average time which wasn't good.

Anyhows, it is best to add another comment instead of editing existing comments so that others would know what were the former results.

If you PR your benchmark I will run it before the 1.6.2 release (but after 1.6.1 will be released)